Validations update customer

+++ src/main/java/NTTDATA/AFP/model/bo/ValidateAfp.java
@@ -7,5 +7,5 @@ import lombok.Data;
 @AllArgsConstructor
 public class ValidateAfp {
     private String dni;
-    private int idAfp;
+    private long idAfp;
 }
